Http exception cookies support (#5197)
