+++ CHANGES/4277.feature
@@ -1,0 +1,1 @@
+Add set_cookie and del_cookie methods to HTTPException
