+++ aiohttp/web_protocol.py
@@ -460,6 +460,7 @@ class RequestHandler(BaseProtocol):
             resp = Response(
                 status=exc.status, reason=exc.reason, text=exc.text, headers=exc.headers
             )
+            resp._cookies = exc._cookies
             reset = await self.finish_response(request, resp, start_time)
         except asyncio.CancelledError:
             raise
