+++ tests/test_web_response.py
@@ -618,93 +618,6 @@ def test_force_close() -> None:
     assert resp.keep_alive is False
 
 
-def test_response_cookies() -> None:
-    resp = StreamResponse()
-
-    assert resp.cookies == {}
-    assert str(resp.cookies) == ""
-
-    resp.set_cookie("name", "value")
-    assert str(resp.cookies) == "Set-Cookie: name=value; Path=/"
-    resp.set_cookie("name", "other_value")
-    assert str(resp.cookies) == "Set-Cookie: name=other_value; Path=/"
-
-    resp.cookies["name"] = "another_other_value"
-    resp.cookies["name"]["max-age"] = 10
-    assert (
-        str(resp.cookies) == "Set-Cookie: name=another_other_value; Max-Age=10; Path=/"
-    )
-
-    resp.del_cookie("name")
-    expected = (
-        'Set-Cookie: name=("")?; '
-        "expires=Thu, 01 Jan 1970 00:00:00 GMT; Max-Age=0; Path=/"
-    )
-    assert Matches(expected) == str(resp.cookies)
-
-    resp.set_cookie("name", "value", domain="local.host")
-    expected = "Set-Cookie: name=value; Domain=local.host; Path=/"
-    assert str(resp.cookies) == expected
-
-
-def test_response_cookie_path() -> None:
-    resp = StreamResponse()
-
-    assert resp.cookies == {}
-
-    resp.set_cookie("name", "value", path="/some/path")
-    assert str(resp.cookies) == "Set-Cookie: name=value; Path=/some/path"
-    resp.set_cookie("name", "value", expires="123")
-    assert str(resp.cookies) == "Set-Cookie: name=value; expires=123; Path=/"
-    resp.set_cookie(
-        "name",
-        "value",
-        domain="example.com",
-        path="/home",
-        expires="123",
-        max_age="10",
-        secure=True,
-        httponly=True,
-        version="2.0",
-        samesite="lax",
-    )
-    assert (
-        str(resp.cookies).lower() == "set-cookie: name=value; "
-        "domain=example.com; "
-        "expires=123; "
-        "httponly; "
-        "max-age=10; "
-        "path=/home; "
-        "samesite=lax; "
-        "secure; "
-        "version=2.0"
-    )
-
-
-def test_response_cookie__issue_del_cookie() -> None:
-    resp = StreamResponse()
-
-    assert resp.cookies == {}
-    assert str(resp.cookies) == ""
-
-    resp.del_cookie("name")
-    expected = (
-        'Set-Cookie: name=("")?; '
-        "expires=Thu, 01 Jan 1970 00:00:00 GMT; Max-Age=0; Path=/"
-    )
-    assert Matches(expected) == str(resp.cookies)
-
-
-def test_cookie_set_after_del() -> None:
-    resp = StreamResponse()
-
-    resp.del_cookie("name")
-    resp.set_cookie("name", "val")
-    # check for Max-Age dropped
-    expected = "Set-Cookie: name=val; Path=/"
-    assert str(resp.cookies) == expected
-
-
 def test_set_status_with_reason() -> None:
     resp = StreamResponse()
